Create README.md
A user is prompted to put in the "map" width and height, the number of people, the number of people who are sick, the number of days for a sick man to die, and the chances of a person in contact of a man sick to get sick. Then the game goes on until either all are dead or all that are sick are dead.